refactor(core): 将后端使用分层架构进行重构，区分出业务层，持久化层，模型层等
