+++ src-tauri/src/app/mod.rs
@@ -1,0 +1,2 @@
+pub mod state;
+pub mod logging;
