+++ src-tauri/src/clients/mod.rs
@@ -1,0 +1,1 @@
+pub mod ollama;
