+++ src-tauri/src/commands/db.rs
@@ -1,7 +1,8 @@
 use tauri::State;
-use crate::AppState;
-use crate::db::{self, TokenStat};
-use crate::error::{AppResult, AppError};
+use crate::app::state::AppState;
+use crate::repositories::stats;
+use crate::domain::models::TokenStat;
+use crate::domain::error::{AppResult, AppError};
 use log::{debug, warn};
 
 /// Command to retrieve token usage statistics from the database.
@@ -9,7 +10,7 @@ use log::{debug, warn};
 pub fn get_token_stats(state: State<'_, AppState>) -> AppResult<Vec<TokenStat>> {
     debug!("Fetching token statistics");
     let conn = state.db.lock().map_err(|e| AppError::Unknown(e.to_string()))?;
-    db::get_aggregated_stats(&conn)
+    stats::get_aggregated_stats(&conn)
 }
 
 /// Command to record token usage in the database.
@@ -23,7 +24,7 @@ pub fn record_tokens(
 ) -> AppResult<()> {
     debug!("Recording token usage - Model: {}, Date: {}, Prompt: {}, Completion: {}", model, date, prompt, completion);
     let conn = state.db.lock().map_err(|e| AppError::Unknown(e.to_string()))?;
-    db::record_tokens(&conn, &date, prompt, completion, &model)
+    stats::record_tokens(&conn, &date, prompt, completion, &model)
 }
 
 /// Command to clear all application data (currently just token statistics).
@@ -31,5 +32,5 @@ pub fn record_tokens(
 pub fn clear_all_data(state: State<'_, AppState>) -> AppResult<()> {
     warn!("Clearing all token statistics database");
     let conn = state.db.lock().map_err(|e| AppError::Unknown(e.to_string()))?;
-    db::clear_database(&conn)
+    stats::clear_database(&conn)
 }
