+++ src-tauri/src/domain/error.rs
